requirements.txt
asgiref==3.5.2
certifi 
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1

Django==4.0.5

django-cors-headers==3.13.0

django-rest-authtoken
django-rest-framework
django-timezone-field
djangorestframework
djangorestframework-simplejwt

gunicorn==20.1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
ipython==8.4.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.29
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_10c245a1-1c81-453c-8b70-3e944787a8d9mo1xhesu/croots/recipe/numpy_and_numpy_base_1653915529188/work
odfpy==1.4.1
opencv-contrib-python==4.5.5.62
openpyxl==3.0.10
packaging==21.3
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.0
Pillow==9.2.0
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==4.21.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rsa==4.8

sqlparse==0.4.2
stack-data==0.3.0
starkbank-ecdsa==2.0.3
starlette==0.19.1
stripe==3.5.0
tablib==3.2.1
traitlets==5.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0

xlwt==1.3.0
